Support fractional stock quantities
- Migrate current_stock, minimum_stock, and movement quantity to
  decimal(12,2); cast to float so values display cleanly (10, 2.5).
- InventoryService stock in/out/adjust accept floats; ProjectService
  deducts/returns exact decimal material amounts (no more rounding up).
- Validation switched integer -> numeric (min 0.01 for movements); stock
  inputs gain step="0.01". Project BOM shortage math uses decimals.
- Add FractionalStockTest (2.25 in, 0.75 out -> 12.00). Suite: 27 passed.

+++ inventory-system/app/Models/InventoryItem.php
@@ -19,8 +19,8 @@ class InventoryItem extends Model
     ];
 
     protected $casts = [
-        'current_stock' => 'integer',
-        'minimum_stock' => 'integer',
+        'current_stock' => 'float',
+        'minimum_stock' => 'float',
         'unit_cost' => 'decimal:2',
     ];
